refactor(core/rust): separate lerp and animations
